# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App API
  description: REST API for the Todo App — create, read, update, and delete todos.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local dev server

tags:
  - name: Todos

paths:
  /todos:
    get:
      summary: List all todos
      tags: [Todos]
      responses:
        "200":
          description: All todos, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"

    post:
      summary: Create a todo
      tags: [Todos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoInput"
            example:
              title: Buy groceries
              description: Milk, eggs, bread
              done: false
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"

  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Todo UUID

    get:
      summary: Get a todo
      tags: [Todos]
      responses:
        "200":
          description: The todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update a todo
      tags: [Todos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoInput"
            example:
              done: true
      responses:
        "200":
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a todo
      tags: [Todos]
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

components:
  schemas:
    Todo:
      type: object
      required: [id, title, done, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          maxLength: 200
          example: Buy groceries
        description:
          type: string
          example: Milk, eggs, bread
        done:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateTodoInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries
        description:
          type: string
          example: Milk, eggs, bread
        done:
          type: boolean
          example: false

    UpdateTodoInput:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        done:
          type: boolean

    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: "`title` is required and must be a non-empty string"

  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Todo not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"